move close_on_error into the only file it's used in
(and update tests)
