+++ trio/tests/test_highlevel_open_tcp_stream.py
@@ -6,27 +6,9 @@ import trio
 from trio.socket import AF_INET, AF_INET6, SOCK_STREAM, IPPROTO_TCP
 from trio._highlevel_open_tcp_stream import (
     reorder_for_rfc_6555_section_5_4,
-    close_on_error,
     open_tcp_stream,
     format_host_port,
 )
-
-
-def test_close_on_error():
-    class CloseMe:
-        closed = False
-
-        def close(self):
-            self.closed = True
-
-    with close_on_error(CloseMe()) as c:
-        pass
-    assert not c.closed
-
-    with pytest.raises(RuntimeError):
-        with close_on_error(CloseMe()) as c:
-            raise RuntimeError
-    assert c.closed
 
 
 def test_reorder_for_rfc_6555_section_5_4():
